change module name DdrPublic to Ddr::Public; fix unrelated problem in spec test; closes #16
