+++ spec/views/catalog/_show_default.html.erb_spec.rb
@@ -8,6 +8,7 @@ RSpec.describe 'catalog/_show_default.html.erb', type: :view do
           'content_size_human_ssim'=>'66.5 MB',
           'read_access_person_ssim'=>['user2'],
           'role_assignments__downloader_ssim'=>'user2',
+          'workflow_state_ssi'=>'published',
           'object_profile_ssm'=>['{"datastreams":{"content":{"dsLabel":"image10.tif","dsVersionID":"content.0","dsCreateDate":"2014-10-22T17:30:02Z","dsState":"A","dsMIME":"image/tiff","dsFormatURI":null,"dsControlGroup":"M","dsSize":69742260,"dsVersionable":true,"dsInfoType":null,"dsLocation":"changeme:10+content+content.0","dsLocationType":"INTERNAL_ID","dsChecksumType":"SHA-256","dsChecksum":"b9eb20b6fb4a27d6bf478bdefb25538bea95740bdf48471ec360d25af622a911"}}}']
           }) }
